requirements.txt
streamlit
PyMuPDF
pyhanko
cryptography
Pillow
